+++ apkInspector/axml.py
@@ -1,10 +1,17 @@
+import logging
 import struct
+
+logging.basicConfig(
+    level=logging.DEBUG,
+    format='%(asctime)s - %(levelname)s - %(filename)s:%(lineno)d -> %(funcName)s : %(message)s'
+)
 
 
 class ResChunkHeader:
     """
-    It is the starting header of the AndroidManifest.xml file.
-    Although the header specification states to start with \x03\x00 it is not validate by Android itself.
+    Chunk header used throughout the axml.
+    This header is essential as it contains information about the header size but also the total size of the chunk
+    the header belongs to.
     """
 
     def __init__(self, header_type, header_size, total_size):
@@ -13,7 +20,7 @@ class ResChunkHeader:
         self.total_size = total_size
 
     @classmethod
-    def from_file(cls, file):
+    def parse(cls, file):
         """
         Read the header type (2 bytes), header size (2 bytes), and entry size (4 bytes)
         :param file: the xml file e.g. with open('/path/AndroidManifest.xml', 'rb') as file:
@@ -44,13 +51,13 @@ class ResStringPoolHeader:
         self.styles_start = styles_start
 
     @classmethod
-    def from_file(cls, file):
+    def parse(cls, file):
         """
         Read and parse ResStringPoolHeader from the file
         :param file: the xml file right after the header has been read.
         :return: Returns an instance of itself
         """
-        header = ResChunkHeader.from_file(file)
+        header = ResChunkHeader.parse(file)
         string_pool_header_data = file.read(20)
         string_count, style_count, flags, strings_start, styles_start = struct.unpack('<IIIII', string_pool_header_data)
         return cls(header.type, header.header_size, header.total_size, string_count, style_count, flags, strings_start,
@@ -62,6 +69,7 @@ class StringPoolType:
     The stringPool class which contains the header defined before: ResStringPoolHeader
     along with the string offsets and the string data.
     """
+
     def __init__(self, header_type, header_size, total_size, string_count, style_count, flags, strings_start,
                  styles_start, string_offsets, strdata):
         self.header = ResStringPoolHeader(header_type, header_size, total_size, string_count, style_count, flags,
@@ -86,7 +94,7 @@ class StringPoolType:
             string_offsets.append(struct.unpack('<I', file.read(4))[0])
         # sanity check as after reading the last string we should be at the end offset as calculated
         if file.tell() != end_absolute_offset:
-            print(
+            logging.warning(
                 f"Current file read:{file.tell()} is not as expected to the start of the stringData:{end_absolute_offset})")
         return string_offsets
 
@@ -140,22 +148,25 @@ class StringPoolType:
         return strings
 
     @classmethod
-    def from_file(cls, file):
+    def parse(cls, file):
         """
         Handle the string pool
         :param file: the xml file right after the file header is read
         :return: Returns an instance of itself
         """
-        string_pool_header = ResStringPoolHeader.from_file(file)
+        string_pool_header = ResStringPoolHeader.parse(file)
         size_of_strings_offsets = string_pool_header.strings_start - 28
         # it should be divisible by 4, as 4 bytes are per offset, so we can get accurately the # of strings
         num_of_strings = size_of_strings_offsets // 4
         if not (size_of_strings_offsets / 4).is_integer():
-            print(f"The number of strings in the string pool is not a integer number.")
+            logging.warning(f"The number of strings in the string pool is not a integer number.")
         string_offsets = cls.read_string_offsets(file, num_of_strings, string_pool_header.strings_start + 8)
         is_utf8 = bool(string_pool_header.flags & (1 << 8))
         string_data = cls.read_strings(file, string_offsets, string_pool_header.strings_start, is_utf8)
         file.read(2)  # the +2 is to account for the null bytes after the last strPool element
+        cur_pos = file.tell()
+        if file.read(2) == b'\x80\x01':
+            file.seek(cur_pos)
         return cls(
             string_pool_header.header.type,
             string_pool_header.header.header_size,
@@ -170,50 +181,52 @@ class StringPoolType:
         )
 
 
-class XmlResourceMapType(ResChunkHeader):
+class XmlResourceMapType:
     """
     Resource map class, with the header and the resource IDs
     """
+
     def __init__(self, header_type, header_size, total_size, resids):
-        super().__init__(header_type, header_size, total_size)
+        self.header = ResChunkHeader(header_type, header_size, total_size)
         self.resids = resids
 
     @classmethod
-    def parse(cls, file, header: ResChunkHeader):
+    def parse(cls, file):
         """
         Parse the resourse map and get the resource IDs
         :param file: the xml file right after the string pool is read
-        :param header: the header ResChunkHeader
         :return: Returns an instance of itself
         """
+        header = ResChunkHeader.parse(file)
         num_resids = (header.total_size - header.header_size) // 4
         resids = [struct.unpack('<I', file.read(4))[0] for _ in range(num_resids)]
 
         return cls(header.type, header.header_size, header.total_size, resids)
 
 
-class XmlStartNamespace(ResChunkHeader):
+class XmlStartNamespace:
     """
     The actual start of the xml, after this the elements of the xml will be found.
     """
+
     def __init__(self, header_type, header_size, total_size, ext):
-        super().__init__(header_type, header_size, total_size)
-        self.ext = ext
+        self.header = ResChunkHeader(header_type, header_size, total_size)
+        self.ext = ext  # [prefix_index, uri_index]
 
     @classmethod
     def parse(cls, file, header: ResChunkHeader):
         num_exts = (header.total_size - header.header_size) // 4
         ext = [struct.unpack('<I', file.read(4))[0] for _ in range(num_exts)]
-
         return cls(header.type, header.header_size, header.total_size, ext)
 
 
-class XmlEndNamespace(ResChunkHeader):
+class XmlEndNamespace:
     """
     Indicator for the end of the xml file.
     """
+
     def __init__(self, header_type, header_size, total_size, prefix_namespace_index, uri_index):
-        super().__init__(header_type, header_size, total_size)
+        self.header = ResChunkHeader(header_type, header_size, total_size)
         self.prefix_namespace_index = prefix_namespace_index
         self.uri_index = uri_index
 
@@ -229,6 +242,7 @@ class XmlAttributeElement:
     """
     The attributes within each element within the xml, should be described by this class
     """
+
     def __init__(self, full_namespace_index, name_index, raw_value_index, typed_value_size, typed_value_res0,
                  typed_value_datatype, typed_value_data):
         self.full_namespace_index = full_namespace_index
@@ -255,13 +269,14 @@ class XmlAttributeElement:
         return attrs
 
 
-class XmlStartElement(ResChunkHeader):
+class XmlStartElement:
     """
     The starting point of an element, its attributes are described by XmlAttributeElement
     The attrext contains information about the element including the attribute count.
     """
+
     def __init__(self, header_type, header_size, total_size, attrext, attributes):
-        super().__init__(header_type, header_size, total_size)
+        self.header = ResChunkHeader(header_type, header_size, total_size)
         self.attrext = attrext
         self.attributes = attributes
 
@@ -276,12 +291,13 @@ class XmlStartElement(ResChunkHeader):
         return cls(header.type, header.header_size, header.total_size, attrext, attributes)
 
 
-class XmlEndElement(ResChunkHeader):
+class XmlEndElement:
     """
     The end of an element, where the attrext contains the necessary information on which element it ends.
     """
+
     def __init__(self, header_type, header_size, total_size, attrext):
-        super().__init__(header_type, header_size, total_size)
+        self.header = ResChunkHeader(header_type, header_size, total_size)
         self.attrext = attrext
 
     @classmethod
@@ -291,8 +307,53 @@ class XmlEndElement(ResChunkHeader):
         return cls(header.type, header.header_size, header.total_size, attrext)
 
 
-def process_xml_resource_map(file, chunk_header: ResChunkHeader):
-    return XmlResourceMapType.parse(file, chunk_header)
+class XmlcDataElement:
+    """
+    A CDATA section
+    https://developer.android.com/reference/org/w3c/dom/CDATASection
+    """
+
+    def __init__(self, header_type, header_size, total_size, data_index, typed_value_size, typed_value_res0,
+                 typed_value_datatype, typed_value_data):
+        self.header = ResChunkHeader(header_type, header_size, total_size)
+        self.data_index = data_index
+        self.typed_value_size = typed_value_size
+        self.typed_value_res0 = typed_value_res0
+        self.typed_value_datatype = typed_value_datatype
+        self.typed_value_data = typed_value_data
+
+    @classmethod
+    def parse(cls, file, header: ResChunkHeader):
+        data_index = struct.unpack('<I', file.read(4))
+        typed_value_size = struct.unpack('<H', file.read(2))[0]
+        typed_value_res0 = struct.unpack('<B', file.read(1))[0]
+        typed_value_datatype = struct.unpack('<B', file.read(1))[0]
+        typed_value_data = struct.unpack('<I', file.read(4))[0]
+        return cls(header.type, header.header_size, header.total_size, data_index, typed_value_size, typed_value_res0,
+                   typed_value_datatype, typed_value_data)
+
+
+class ManifestStruct:
+    """
+    A class to represent the AndroidManifest as a composition
+    """
+    def __init__(self, header: ResChunkHeader, string_pool: StringPoolType, resource_map: XmlResourceMapType, elements):
+        self.header = header
+        self.string_pool = string_pool
+        self.resource_map = resource_map
+        self.elements = elements
+
+    def get_manifest(self):
+        manifest = create_manifest(self.elements, self.string_pool.strdata)
+        return manifest
+
+    @classmethod
+    def parse(cls, file):
+        header = ResChunkHeader.parse(file)
+        string_pool = StringPoolType.parse(file)
+        resource_map = XmlResourceMapType.parse(file)
+        elements = process_elements(file)[0]
+        return cls(header, string_pool, resource_map, elements)
 
 
 def process_xml_start_namespace(file, chunk_header: ResChunkHeader):
@@ -311,12 +372,16 @@ def process_xml_end_element(file, chunk_header: ResChunkHeader):
     return XmlEndElement.parse(file, chunk_header)
 
 
+def process_cdata(file, chunk_header: ResChunkHeader):
+    return XmlcDataElement.parse(file, chunk_header)
+
+
 chunk_type_handlers = {
-    '0x180': process_xml_resource_map,      # RES_XML_RESOURCE_MAP_TYPE
     '0x100': process_xml_start_namespace,   # RES_XML_START_NAMESPACE_TYPE
+    '0x101': process_xml_end_namespace,     # RES_XML_END_NAMESPACE_TYPE
     '0x102': process_xml_start_element,     # RES_XML_START_ELEMENT_TYPE
     '0x103': process_xml_end_element,       # RES_XML_END_ELEMENT_TYPE
-    '0x101': process_xml_end_namespace,      # RES_XML_END_NAMESPACE_TYPE
+    '0x104': process_cdata,                 # RES_XML_CDATA_TYPE
 }
 
 
@@ -328,7 +393,7 @@ def parse_next_header(file):
     :param file: the xml file that is being read
     :return: Dispatches to the appropriate processing method for each chunk type.
     """
-    chunk_header = ResChunkHeader.from_file(file)
+    chunk_header = ResChunkHeader.parse(file)
     if chunk_header is None:  # end of file
         return None
     if chunk_header.header_size > 8:
@@ -341,17 +406,15 @@ def parse_next_header(file):
         raise NotImplementedError(f"Unsupported chunk type: {chunk_type}")
 
 
-def process_headers(file):
+def process_elements(file):
     """
-    It starts processing the remaining headers after the string pool chunk.
-    Takes into account that the resource map, the start namespace and end namespace chunks are only to be found once
-    within the file.
-    :param file: the xml file after the string pool chunk
+    It starts processing the remaining chunks after the resource map chunk.
+    :param file: the xml file read right after the resource map chunk
     :return: Returns all the elements found as their corresponding classes and whether dummy data were found in between
     """
     elements = []
     dummy = 0
-    possible_headers = {b'\x80\x01', b'\x00\x01', b'\x02\x01', b'\x03\x01', b'\x01\x01'}
+    possible_headers = {b'\x00\x01', b'\x01\x01', b'\x02\x01', b'\x03\x01', b'\x04\x01'}
     while True:
         # Parse the next header
         cur_pos = file.tell()
@@ -366,75 +429,83 @@ def process_headers(file):
             continue
         chunk_type = parse_next_header(file)
         elements.append(chunk_type)
-        if check in {b'\x80\x01', b'\x00\x01', b'\x01\x01'}:
-            possible_headers.remove(check)
     return elements, dummy
+
+
+def process_attributes(attributes, string_data, ns_dict):
+    """
+    Helps in processing the attributes found in each element of the axml
+    :param attributes: the attributes of an XmlStartElement object
+    :param string_data: the string data list from the String Pool
+    :param ns_dict: a namespace dictionary based on the XmlStartNamespace elements found
+    :return: returns a string of all the attributes with their values
+    """
+    attribute_list = []
+    for attr in attributes:
+        name = string_data[attr.name_index]
+        if attr.typed_value_datatype == 1:  # reference type
+            value = f"@{attr.typed_value_data}"
+        elif attr.typed_value_datatype == 3:  # string type
+            try:
+                value = string_data[attr.typed_value_data]
+            except:
+                value = attr.typed_value_data
+        elif attr.typed_value_datatype == 16:  # integer type
+            value = str(attr.typed_value_data)
+        elif attr.typed_value_datatype == 17:  # int-hex type
+            value = f"{attr.typed_value_data} ({hex(attr.typed_value_data)})"
+        elif attr.typed_value_datatype == 18:  # boolean type
+            value = "true" if bool(attr.typed_value_data) else "false"
+        elif attr.typed_value_datatype == 0:  # null, used for CData
+            return name
+        else:
+            logging.error(f"An unknown datatype came up: {attr.typed_value_datatype}")
+        if attr.full_namespace_index < len(string_data):
+            namespace = string_data[attr.full_namespace_index]
+            attribute_list.append(f'{ns_dict[namespace]}:{name}="{value}"')
+        else:
+            attribute_list.append(f'{name}="{value}"')
+
+    return ' '.join(attribute_list)
 
 
 def create_manifest(elements, string_data):
     """
-    Method to go over all elements and attempt to create the AndroidManifest.xml file.
-    :param elements: Elements as retrieved from process_headers()
-    :param string_data: The string data contained within the string pool
-    :return: Returns the string of the AndroidManifest.xml
+
+    :param elements:
+    :param string_data:
+    :return:
     """
-    android_manifest_xml = '<?xml version="1.0" encoding="utf-8"?>\n'
-    namespace = ""
+    android_manifest_xml = []
+    namespaces = []
+    ns_dict = {}
+
     for element in elements:
         if isinstance(element, XmlStartNamespace):
-            namespace = f'xmlns:{string_data[element.ext[0]]}="{string_data[element.ext[1]]}"'
+            namespaces.append(f'xmlns:{string_data[element.ext[0]]}="{string_data[element.ext[1]]}"')
+            ns_dict[string_data[element.ext[1]]] = string_data[element.ext[0]]
         elif isinstance(element, XmlStartElement):
-            ln_list = []
-            for attr in element.attributes:
-                if attr.typed_value_datatype == 1:  # reference type
-                    name = string_data[attr.name_index]
-                    value = f"@{attr.typed_value_data}"
-                elif attr.typed_value_datatype == 3:  # string type
-                    name = string_data[attr.name_index]
-                    try:
-                        value = string_data[attr.typed_value_data]
-                    except:
-                        value = attr.typed_value_data
-                elif attr.typed_value_datatype == 16:  # integer type
-                    name = string_data[attr.name_index]
-                    value = attr.typed_value_data
-                elif attr.typed_value_datatype == 17:  # int-hex type
-                    name = string_data[attr.name_index]
-                    value = f"{attr.typed_value_data} ({hex(attr.typed_value_data)})"
-                elif attr.typed_value_datatype == 18:  # boolean type
-                    name = string_data[attr.name_index]
-                    value = "true" if bool(attr.typed_value_data) else "false"
-                else:
-                    print(f"An unknown datatype came up: {attr.typed_value_datatype}")
-                ln_list.append(f'android:{name}="{value}"')
-            if ln_list:
-                if string_data[element.attrext[1]] == "manifest":
-                    android_manifest_xml += f"<{string_data[element.attrext[1]]} {namespace} {' '.join(ln_list)}>\n"
-                else:
-                    android_manifest_xml += f"<{string_data[element.attrext[1]]} {' '.join(ln_list)}>\n"
+            attributes = process_attributes(element.attributes, string_data, ns_dict)
+            if string_data[element.attrext[1]] == 'manifest':
+                tag_line = f"<{string_data[element.attrext[1]]} {' '.join(namespaces)} {attributes}>\n" if attributes else f"<{string_data[element.attrext[1]]}>\n"
             else:
-                android_manifest_xml += f"<{string_data[element.attrext[1]]}>\n"
+                tag_line = f"<{string_data[element.attrext[1]]} {attributes}>\n" if attributes else f"<{string_data[element.attrext[1]]}>\n"
+            android_manifest_xml.append(tag_line)
+        elif isinstance(element, XmlcDataElement):
+            if android_manifest_xml[-1][-1] == '\n':
+                android_manifest_xml[-1] = android_manifest_xml[-1].replace('\n', string_data[element.data_index[0]])
         elif isinstance(element, XmlEndElement):
             name = string_data[element.attrext[1]]
-            if name == "manifest":
-                android_manifest_xml += f"</{string_data[element.attrext[1]]}>"
-            else:
-                android_manifest_xml += f"</{string_data[element.attrext[1]]}>\n"
-    return android_manifest_xml
+            closing_tag = f"</{name}>" if name == "manifest" else f"</{name}>\n"
+            android_manifest_xml.append(closing_tag)
+    return ''.join(android_manifest_xml)
 
 
-def get_manifest(file_like_object):
+def get_manifest(raw_manifest):
     """
-    Method to return the AndroidManifest file as created by create_manifest()
-    :param file_like_object: expects the encoded AndroidManifest.xml file as a file-like object
+    Method to directly return the AndroidManifest file as created by create_manifest()
+    :param raw_manifest: expects the encoded AndroidManifest.xml file as a file-like object
     :return: returns the decoded AndroidManifest file
     """
-    ResChunkHeader.from_file(file_like_object)
-    string_pool = StringPoolType.from_file(file_like_object)
-    string_data = string_pool.strdata
-    elements = process_headers(file_like_object)[0]
-    manifest = create_manifest(elements, string_data)
-    return manifest
-
-
-
+    manifest_object = ManifestStruct.parse(raw_manifest)
+    return manifest_object.get_manifest()